# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CollectionResponse:
    properties:
      cover_image:
        type: string
      description:
        type: string
      id:
        type: string
      is_custom:
        type: boolean
      is_public:
        type: boolean
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  handler.CreateCollectionInput:
    properties:
      cover_image:
        type: string
      description:
        type: string
      is_custom:
        type: boolean
      is_public:
        type: boolean
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.PaginatedCollectionsResponse:
    properties:
      collections:
        items:
          $ref: '#/definitions/handler.CollectionResponse'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handler.signInInput:
    properties:
      mail:
        type: string
      password:
        type: string
    required:
    - mail
    - password
    type: object
  handler.signInResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  handler.signUpInput:
    properties:
      mail:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - mail
    - name
    - password
    type: object
  models.UserResponse:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      mail:
        type: string
      name:
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Memoria API
  version: "1.0"
paths:
  /auth/callback/github:
    get:
      description: Обрабатывает callback от GitHub после аутентификации
      parameters:
      - description: Код авторизации от GitHub
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/handler.signInResponse'
        "400":
          description: Отсутствует код авторизации
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: OAuth callback от GitHub
      tags:
      - auth
...
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по email и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/handler.signInResponse'
        "400":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /collections:
    get:
      description: Возвращает пагинированный список коллекций пользователя
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Кол-во элементов на странице
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handler.PaginatedCollectionsResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить список коллекций
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Создает новую коллекцию для пользователя
      parameters:
      - description: Данные для создания коллекции
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCollectionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Коллекция создана
          schema:
            $ref: '#/definitions/handler.CollectionResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать коллекцию
      tags:
      - collections
...
securityDefinitions:
  apiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"